Put the new logic into the existing ControlOfTheDream.tsx that was not working correctly
