+++ src/analysis/retail/druid/balance/CombatLogParser.ts
@@ -30,7 +30,6 @@ import CelestialAlignment from 'analysis/retail/druid/balance/modules/spells/Cel
 import ControlOfTheDream from 'analysis/retail/druid/shared/spells/ControlOfTheDream';
 import CastLinkNormalizer from 'analysis/retail/druid/balance/normalizers/CastLinkNormalizer';
 import Lunation from 'analysis/retail/druid/shared/spells/Lunation';
-import SpellUsable from 'analysis/retail/druid/balance/modules/features/SpellUsable';
 import Haste from 'analysis/retail/druid/balance/modules/core/Haste';
 
 class CombatLogParser extends MainCombatLogParser {
@@ -61,7 +60,7 @@ class CombatLogParser extends MainCombatLogParser {
     buffs: Buffs,
     dotUptimes: DotUptimes,
     eclipse: Eclipse,
-    spellUsable: SpellUsable,
+
     //Talents
     twinMoons: TwinMoons,
     starlord: Starlord,
@@ -72,6 +71,7 @@ class CombatLogParser extends MainCombatLogParser {
     newMoon: NewMoon,
     wildMushroom: WildMushroom,
     celestialAlignment: CelestialAlignment,
+
     //Hero Talents
     lunation: Lunation,
     controlOfTheDream: ControlOfTheDream,
