+++ src/analysis/retail/druid/balance/modules/Abilities.tsx
@@ -85,7 +85,9 @@ class Abilities extends CoreAbilities {
           combatant.hasTalent(TALENTS_DRUID.INCARNATION_CHOSEN_OF_ELUNE_TALENT),
         castEfficiency: {
           suggestion: true,
-          recommendedEfficiency: 0.8,
+          recommendedEfficiency: 0.9,
+          averageIssueEfficiency: 0.8,
+          majorIssueEfficiency: 0.7,
         },
         charges: combatant.hasTalent(TALENTS_DRUID.WHIRLING_STARS_TALENT) ? 2 : 1,
         gcd: combatant.hasTalent(TALENTS_DRUID.ORBITAL_STRIKE_TALENT) ? { base: 1500 } : null,
@@ -101,7 +103,9 @@ class Abilities extends CoreAbilities {
         },
         castEfficiency: {
           suggestion: true,
-          recommendedEfficiency: 0.8,
+          recommendedEfficiency: 0.9,
+          averageIssueEfficiency: 0.8,
+          majorIssueEfficiency: 0.7,
         },
         timelineSortIndex: 10,
       },
