+++ src/analysis/retail/druid/shared/spells/ControlOfTheDream.tsx
@@ -5,29 +5,36 @@ import Events, { UpdateSpellUsableEvent, UpdateSpellUsableType } from 'parser/co
 import SpellUsable from 'parser/shared/modules/SpellUsable';
 import Abilities from 'parser/core/modules/Abilities';
 import Statistic from 'parser/ui/Statistic';
-import { Fragment } from 'react';
 import STATISTIC_ORDER from 'parser/ui/STATISTIC_ORDER';
 import BoringSpellValueText from 'parser/ui/BoringSpellValueText';
 import STATISTIC_CATEGORY from 'parser/ui/STATISTIC_CATEGORY';
 import { SpellIcon } from 'interface';
+import SPELLS from 'common/SPELLS';
+
+const DEBUG = false;
 
 const MAJOR_SPELLS = [
   TALENTS_DRUID.FORCE_OF_NATURE_TALENT,
-  TALENTS_DRUID.CELESTIAL_ALIGNMENT_TALENT,
-  TALENTS_DRUID.INCARNATION_CHOSEN_OF_ELUNE_TALENT,
-  TALENTS_DRUID.CONVOKE_THE_SPIRITS_TALENT,
-  TALENTS_DRUID.NATURES_SWIFTNESS_TALENT,
+  SPELLS.INCARNATION_ORBITAL_STRIKE,
+  SPELLS.INCARNATION_CHOSEN_OF_ELUNE,
+  SPELLS.CELESTIAL_ALIGNMENT_ORBITAL_STRIKE,
+  SPELLS.CELESTIAL_ALIGNMENT,
+  SPELLS.CONVOKE_SPIRITS,
+  SPELLS.NATURES_SWIFTNESS,
   TALENTS_DRUID.INCARNATION_TREE_OF_LIFE_TALENT,
 ];
 
 const MAX_CDR = 15_000;
 
-/**
- * **Control of the Dream**
+/** **Control of the Dream**
+ *
+ * https://www.wowhead.com/spell=434249/control-of-the-dream
+ *
  * Keeper of the Grove Hero Talent
  *
- * Time elapsed while your major abilities are available to be used is subtracted from that
- * ability's cooldown after the next time you use it, up to 15 seconds.
+ * (as-of 12.0.5)
+ * Time elapsed while your major abilities are available to be used or at maximum charges
+ * is subtracted from that ability's cooldown after the next time you use it, up to 15 seconds.
  * Balance: Force of Nature, Celestial Alignment, Incarnation: CoE, Convoke the Spirits
  * Resto: Nature's Swiftness, Incarnation: ToL, Convoke the Spirits
  */
@@ -38,6 +45,17 @@ export default class ControlOfTheDream extends Analyzer.withDependencies({
   /** Info about each 'major abilities' CDR, indexed by spellId */
   cdrSpellInfos: CdrSpellInfo[] = [];
 
+  get totalEarlyCastCount(): number {
+    return this.cdrSpellInfos.reduce((sum, cdrSpellInfo) => sum + cdrSpellInfo.earlyCastCount, 0);
+  }
+
+  get totalEffectiveCdr(): number {
+    return (
+      this.cdrSpellInfos.reduce((sum, cdrSpellInfo) => sum + cdrSpellInfo.totalEffectiveCdr, 0) /
+      1_000
+    );
+  }
+
   constructor(options: Options) {
     super(options);
     this.active = this.selectedCombatant.hasTalent(TALENTS_DRUID.CONTROL_OF_THE_DREAM_TALENT);
@@ -47,14 +65,21 @@ export default class ControlOfTheDream extends Analyzer.withDependencies({
       this.onMajorSpellCdUpdate,
     );
 
-    // the unusual CD behavior requires a custom maxSpells in CastEfficiency
-    // (one that basically ignores this abilities CDR except the first cast, as it evens out in the end)
+    // This Talent does not actually reduce the CD overall; it just gives the player a 15s window
+    // to delay the cast without losing efficiency. But if we wait for 15s before casting again,
+    // then the next CD will be 15s shorter, and that messes up the computed effective cooldown.
+    // Therefore, the default implementation of "maxSpells" in "CastEfficiency",
+    // based on the effective cooldown, is wrong.
+    // Let's ignore this ability's CDR except for the first cast.
     if (this.active) {
       MAJOR_SPELLS.forEach((spell) => {
         const ability = (options.abilities as Abilities).getAbility(spell.id);
+        const abilityMaxCharges = this.deps.abilities.getMaxCharges(spell.id) || 1;
         if (ability) {
           ability.castEfficiency.maxCasts = (cooldown) =>
-            (this.owner.fightDuration + MAX_CDR) / (cooldown * 1_000);
+            (this.owner.fightDuration + MAX_CDR) / (cooldown * 1_000) +
+            // Take into account extra charges (e.g. Incarnation: CoE with Whirling Stars talent)
+            (abilityMaxCharges - 1);
         }
       });
     }
@@ -64,9 +89,9 @@ export default class ControlOfTheDream extends Analyzer.withDependencies({
     const spellId = event.ability.guid;
     if (!this.cdrSpellInfos[spellId]) {
       this.cdrSpellInfos[spellId] = {
-        earlyCasts: 0,
+        earlyCastCount: 0,
         totalEffectiveCdr: 0,
-        cotdCdrPrevCycle: 0,
+        lastCdrApplied: 0,
       };
     }
     const info = this.cdrSpellInfos[spellId];
@@ -74,22 +99,73 @@ export default class ControlOfTheDream extends Analyzer.withDependencies({
     if (event.updateType === UpdateSpellUsableType.BeginCooldown) {
       // Check if this cast benefited from CotD by comparing against when the spell
       // would have been available without CotD's CDR (but with other sources like Cenarius' Guidance)
-      if (info.lastAvailable !== undefined && info.cotdCdrPrevCycle > 0) {
-        const wouldBeAvailableWithoutCotd = info.lastAvailable + info.cotdCdrPrevCycle;
-        const effectiveCdr = Math.max(0, wouldBeAvailableWithoutCotd - this.owner.currentTimestamp);
-        if (effectiveCdr > 0) {
-          info.earlyCasts += 1;
-          info.totalEffectiveCdr += effectiveCdr;
+      if (info.lastAvailable !== undefined && info.lastCdrApplied > 0) {
+        const lastAvailableWithoutCdrApplied = info.lastAvailable + info.lastCdrApplied;
+        const earlyCastTimeGained = Math.max(0, lastAvailableWithoutCdrApplied - event.timestamp);
+        if (earlyCastTimeGained > 0) {
+          info.earlyCastCount += 1;
+          info.totalEffectiveCdr += earlyCastTimeGained;
         }
       }
-      // First cast always gets max CDR
-      const cdr = !info.lastAvailable
-        ? MAX_CDR
-        : Math.min(MAX_CDR, this.owner.currentTimestamp - info.lastAvailable);
-      info.cotdCdrPrevCycle = this.deps.spellUsable.reduceCooldown(spellId, cdr);
+      const effectiveCDR = this.reduceCooldown(info, spellId, event);
+      info.lastCdrApplied = effectiveCDR;
     } else if (event.updateType === UpdateSpellUsableType.EndCooldown) {
-      // a major ability just finished CD, register it
-      info.lastAvailable = this.owner.currentTimestamp;
+      // A major ability just finished CD, register it
+      info.lastAvailable = event.timestamp;
+    }
+  }
+
+  private reduceCooldown(
+    info: CdrSpellInfo,
+    spellId: number,
+    event: UpdateSpellUsableEvent,
+  ): number {
+    // We can discard spells that do not have their max number of charges
+    if (this.hasNotMaximumNumberOfCharges(spellId, event)) {
+      DEBUG &&
+        console.info(
+          `[${event.timestamp}] Cooldown of ${spellId} is not reduced as it's below max number of charges`,
+        );
+      return 0;
+    }
+
+    const cdr = this.getCooldownReductionMs(info, event);
+    return this.deps.spellUsable.reduceCooldown(spellId, cdr);
+  }
+
+  private hasNotMaximumNumberOfCharges(spellId: number, event: UpdateSpellUsableEvent): boolean {
+    const abilityMaxCharges = this.deps.abilities.getMaxCharges(spellId);
+    if (abilityMaxCharges && abilityMaxCharges > 1) {
+      const ability = this.deps.abilities.getAbility(spellId);
+      if (!ability) {
+        return false;
+      }
+
+      const abilityCharges = this.deps.spellUsable.chargesAvailable(spellId);
+      DEBUG &&
+        console.info(`[${event.timestamp}] ${spellId} has ${abilityCharges} remaining charges.`);
+
+      // This method is invoked after the ability has been cast, thus we add 1 charge
+      // to the current number of charges to know how many charges were available before this cast.
+      return abilityCharges + 1 < abilityMaxCharges;
+    } else {
+      // Ability with only 1 charge always has maximum number of charges (1) when cast
+      return false;
+    }
+  }
+
+  private getCooldownReductionMs(info: CdrSpellInfo, event: UpdateSpellUsableEvent): number {
+    if (info.lastAvailable === undefined) {
+      // Assume ability was available for at least the cap duration (15s) pre-pull.
+      // We have no way of knowing exactly how long it was off CD pre-combat,
+      // but this is a reasonnable assumption for most pull scenarios.
+      return MAX_CDR;
+    } else {
+      // Otherwise, we need to compute how long the ability has been off CD
+      const timeElapsedSinceSpellAvailable = event.timestamp - info.lastAvailable;
+
+      // CD reduction is capped.
+      return Math.min(timeElapsedSinceSpellAvailable, MAX_CDR);
     }
   }
 
@@ -99,15 +175,49 @@ export default class ControlOfTheDream extends Analyzer.withDependencies({
         position={STATISTIC_ORDER.CORE(1)}
         category={STATISTIC_CATEGORY.HERO_TALENTS}
         size="flexible"
+        dropdown={
+          <>
+            <table className="table table-condensed">
+              <thead>
+                <tr>
+                  <th>Ability</th>
+                  <th>Early casts</th>
+                  <th>Eff. CDR</th>
+                </tr>
+              </thead>
+              <tbody>
+                {this.cdrSpellInfos.map((cdrInfo, spellId) => (
+                  <tr>
+                    <th>
+                      <SpellIcon spell={spellId} />
+                    </th>
+                    <td>{cdrInfo.earlyCastCount}</td>
+                    <td>{(cdrInfo.totalEffectiveCdr / 1_000).toFixed(0)}s</td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+          </>
+        }
+        tooltip={
+          <>
+            <p>
+              <b>Early Cast:</b> A cast benefiting from this Talent’s CD reduction.
+            </p>
+            <p>
+              <b>Effective CDR:</b> Total CD time saved via Early Casts.
+            </p>
+          </>
+        }
       >
         <BoringSpellValueText spell={TALENTS_DRUID.CONTROL_OF_THE_DREAM_TALENT}>
           <>
-            {this.cdrSpellInfos.map((cdrInfo, spellId) => (
-              <Fragment key={spellId}>
-                <SpellIcon spell={spellId} /> {(cdrInfo.totalEffectiveCdr / 1_000).toFixed(0)}s{' '}
-                <small>eff. CDR</small> / {cdrInfo.earlyCasts} <small>early casts</small>
-              </Fragment>
-            ))}
+            <div>
+              {this.totalEarlyCastCount} <small>Early Casts</small>
+            </div>
+            <div>
+              {this.totalEffectiveCdr.toFixed(0)}s <small>Effective CDR</small>
+            </div>
           </>
         </BoringSpellValueText>
       </Statistic>
@@ -119,9 +229,9 @@ interface CdrSpellInfo {
   /** Timestamp spell last became available (cooldown finished) */
   lastAvailable?: number;
   /** CotD CDR applied during the previous cooldown cycle, in ms */
-  cotdCdrPrevCycle: number;
+  lastCdrApplied: number;
   /** Times spell was cast earlier than would have been possible without CotD */
-  earlyCasts: number;
+  earlyCastCount: number;
   /** Sum of effective 'early cast' CDR, in ms */
   totalEffectiveCdr: number;
 }
